Fix broken RunIt tests
Moved test-data into run-it
