+++ packages/api-explorer/src/test-data/data.ts
@@ -23,30 +23,9 @@
  SOFTWARE.
 
  */
-import { ApiModel } from '@looker/sdk-codegen'
+import { specs } from '@looker/run-it/src/test-data'
 
 import { SpecItems } from '../ApiExplorer'
 import { initDefaultSpecState } from '../reducers'
 
-export const specs: SpecItems = {
-  '3.0': {
-    status: 'stable',
-    specURL: 'https://self-signed.looker.com:19999/api/3.0/swagger.json',
-    specContent: require('../../specs/Looker.3.0.oas.json'),
-  },
-  '3.1': {
-    status: 'current',
-    specURL: 'https://self-signed.looker.com:19999/api/3.1/swagger.json',
-    specContent: require('../../specs/Looker.3.1.oas.json'),
-  },
-  '4.0': {
-    isDefault: true,
-    status: 'experimental',
-    specURL: 'https://self-signed.looker.com:19999/api/4.0/swagger.json',
-    specContent: require('../../specs/Looker.4.0.oas.json'),
-  },
-}
-
-export const specState = initDefaultSpecState(specs)
-
-export const api = ApiModel.fromJson(specs['3.1'].specContent)
+export const specState = initDefaultSpecState(specs as SpecItems)
