+++ packages/run-it/src/test-data/specs.ts
@@ -23,4 +23,20 @@
  SOFTWARE.
 
  */
-export { specState } from './data'
+import { ApiModel } from '@looker/sdk-codegen'
+
+export const specs = {
+  '3.1': {
+    status: 'current',
+    specURL: 'https://self-signed.looker.com:19999/api/3.1/swagger.json',
+    specContent: require('../../../../spec/Looker.3.1.oas.json'),
+  },
+  '4.0': {
+    isDefault: true,
+    status: 'experimental',
+    specURL: 'https://self-signed.looker.com:19999/api/4.0/swagger.json',
+    specContent: require('../../../../spec/Looker.4.0.oas.json'),
+  },
+}
+
+export const api = ApiModel.fromJson(specs['3.1'].specContent)
